update reservation - MAIS probleme global pour editer les dates reste à régler

+++ src/main/java/com/epf/rentmanager/dao/ClientDao.java
@@ -43,7 +43,7 @@ public class ClientDao {
         }
     }
 
-    public long create(Client client) throws DaoException {
+    public int create(Client client) throws DaoException {
         try {
             Connection connection = ConnectionManager.getConnection();
             Statement statement = connection.createStatement();
